Fix some int, and move JuegoVampiros up 1 folder
